+++ workers/unity/Packages/io.improbable.gdk.core/Commands/ICommandDiffDeserializer.cs
@@ -1,0 +1,10 @@
+using Improbable.Worker.CInterop;
+
+namespace Improbable.Gdk.Core.Commands
+{
+    public interface ICommandDiffDeserializer
+    {
+        void AddRequestToDiff(CommandRequestOp op, ViewDiff diff);
+        void AddResponseToDiff(CommandResponseOp op, ViewDiff diff, CommandMetaData commandMetaData);
+    }
+}