+++ workers/unity/Packages/io.improbable.gdk.core/Commands/IComponentCommandSendStorage.cs
@@ -1,0 +1,13 @@
+using System;
+
+namespace Improbable.Gdk.Core.Commands
+{
+    public interface IComponentCommandSendStorage : ICommandSendStorage
+    {
+        uint ComponentId { get; }
+        uint CommandId { get; }
+
+        Type RequestType { get; }
+        Type ResponseType { get; }
+    }
+}