+++ workers/unity/Packages/io.improbable.gdk.core/Commands/WorldCommands/CreateEntity.cs
@@ -80,9 +80,9 @@ namespace Improbable.Gdk.Core.Commands
                 /// <summary>
                 ///     The unique request ID of this command. Will match the request ID in the corresponding request.
                 /// </summary>
-                public readonly long RequestId;
+                public readonly CommandRequestId RequestId;
 
-                internal ReceivedResponse(CreateEntityResponseOp op, Entity sendingEntity, Request req, long requestId)
+                internal ReceivedResponse(CreateEntityResponseOp op, Entity sendingEntity, Request req, CommandRequestId requestId)
                 {
                     SendingEntity = sendingEntity;
                     StatusCode = op.StatusCode;
@@ -96,7 +96,7 @@ namespace Improbable.Gdk.Core.Commands
                         : (EntityId?) null;
                 }
 
-                long IReceivedCommandResponse.RequestId => RequestId;
+                CommandRequestId IReceivedCommandResponse.RequestId => RequestId;
             }
         }
     }